# tcl-web/src/types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ToneCards Lab API",
    "description": "ToneCards Lab API for managing musical language flashcards",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/users/": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Create User",
        "description": "Create a new user account.",
        "operationId": "create_user_api_v1_users__post",
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Email"
            }
          },
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_UserProfileResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "users"
        ],
        "summary": "List Users",
        "description": "List users with optional search.",
        "operationId": "list_users_api_v1_users__get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_List_UserProfileResponse__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/{user_id}": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get User",
        "description": "Get user profile by ID.",
        "operationId": "get_user_api_v1_users__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_UserProfileResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "users"
        ],
        "summary": "Update User",
        "description": "Update user profile.",
        "operationId": "update_user_api_v1_users__user_id__patch",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserProfileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_UserProfileResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "users"
        ],
        "summary": "Delete User",
        "description": "Delete user account (soft delete).",
        "operationId": "delete_user_api_v1_users__user_id__delete",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteAccountRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/email/{email}": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get User By Email",
        "description": "Get user profile by email address.",
        "operationId": "get_user_by_email_api_v1_users_email__email__get",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Email"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_UserProfileResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/files/user/{user_id}": {
      "get": {
        "tags": [
          "files"
        ],
        "summary": "Get User Files",
        "description": "Get all files uploaded by a user.",
        "operationId": "get_user_files_api_v1_files_user__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_UserFilesResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Indicates if the request was successful",
            "default": true
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "description": "Response message or error description"
          },
          "data": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "Response payload"
          },
          "metadata": {
            "$ref": "#/components/schemas/Metadata",
            "description": "Response metadata"
          },
          "errors": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors",
            "description": "Error details if any"
          }
        },
        "type": "object",
        "title": "ApiResponse"
      },
...
      "ApiResponse_List_UserProfileResponse__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Indicates if the request was successful",
            "default": true
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "description": "Response message or error description"
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/UserProfileResponse"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "Response payload"
          },
          "metadata": {
            "$ref": "#/components/schemas/Metadata",
            "description": "Response metadata"
          },
          "errors": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors",
            "description": "Error details if any"
          }
        },
        "type": "object",
        "title": "ApiResponse[List[UserProfileResponse]]"
      },
...
      "ApiResponse_UserFilesResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Indicates if the request was successful",
            "default": true
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "description": "Response message or error description"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserFilesResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Response payload"
          },
          "metadata": {
            "$ref": "#/components/schemas/Metadata",
            "description": "Response metadata"
          },
          "errors": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors",
            "description": "Error details if any"
          }
        },
        "type": "object",
        "title": "ApiResponse[UserFilesResponse]"
      },
      "ApiResponse_UserProfileResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Indicates if the request was successful",
            "default": true
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "description": "Response message or error description"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserProfileResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Response payload"
          },
          "metadata": {
            "$ref": "#/components/schemas/Metadata",
            "description": "Response metadata"
          },
          "errors": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors",
            "description": "Error details if any"
          }
        },
        "type": "object",
        "title": "ApiResponse[UserProfileResponse]"
      },
...
      "DeleteAccountRequest": {
        "properties": {
          "confirmation": {
            "type": "string",
            "title": "Confirmation",
            "description": "Confirmation string (e.g., 'DELETE' or user email)"
          },
          "transfer_decks": {
            "type": "boolean",
            "title": "Transfer Decks",
            "description": "Whether to transfer owned decks to viewers",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "confirmation"
        ],
        "title": "DeleteAccountRequest",
        "description": "Request to delete user account."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Metadata": {
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp",
            "description": "Response timestamp"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id",
            "description": "Unique request identifier"
          },
          "version": {
            "type": "string",
            "title": "Version",
            "description": "API version",
            "default": "1.0"
          },
          "total_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Count",
            "description": "Total number of items (for paginated responses)"
          },
          "page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page",
            "description": "Current page number"
          },
          "page_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Size",
            "description": "Number of items per page"
          },
          "execution_time_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Time Ms",
            "description": "Request execution time in milliseconds"
          }
        },
        "type": "object",
        "title": "Metadata"
      },
...
      "UpdateUserProfileRequest": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "Updated display name"
          },
          "preferences": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferences",
            "description": "Updated user preferences"
          }
        },
        "type": "object",
        "title": "UpdateUserProfileRequest",
        "description": "Request to update user profile."
      },
      "UploadedFile": {
        "properties": {
          "file_id": {
            "type": "string",
            "title": "File Id",
            "description": "UUID of the file"
          },
          "file_url": {
            "type": "string",
            "title": "File Url",
            "description": "URL to access the file"
          },
          "file_name": {
            "type": "string",
            "title": "File Name",
            "description": "Original filename"
          },
          "file_size": {
            "type": "integer",
            "title": "File Size",
            "description": "File size in bytes"
          },
          "mime_type": {
            "type": "string",
            "title": "Mime Type",
            "description": "MIME type"
          },
          "uploaded_at": {
            "type": "string",
            "format": "date-time",
            "title": "Uploaded At",
            "description": "Upload timestamp"
          },
          "deck_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deck Id",
            "description": "Associated deck ID if applicable"
          }
        },
        "type": "object",
        "required": [
          "file_id",
          "file_url",
          "file_name",
          "file_size",
          "mime_type",
          "uploaded_at"
        ],
        "title": "UploadedFile",
        "description": "Uploaded file metadata."
      },
      "UserFilesResponse": {
        "properties": {
          "files": {
            "items": {
              "$ref": "#/components/schemas/UploadedFile"
            },
            "type": "array",
            "title": "Files",
            "description": "List of uploaded files"
          },
          "total_size": {
            "type": "integer",
            "title": "Total Size",
            "description": "Total size of all files in bytes"
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count",
            "description": "Total number of files"
          }
        },
        "type": "object",
        "required": [
          "total_size",
          "total_count"
        ],
        "title": "UserFilesResponse",
        "description": "Response with list of user's uploaded files."
      },
      "UserProfileResponse": {
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Timestamp when resource was created"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "Timestamp when resource was last updated"
          },
          "user_id": {
            "type": "string",
            "title": "User Id",
            "description": "UUID of the user"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "User's email address"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "User's display name"
          },
          "preferences": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferences",
            "description": "User preferences/settings"
          },
          "owned_decks_count": {
            "type": "integer",
            "title": "Owned Decks Count",
            "description": "Number of decks owned by user",
            "default": 0
          },
          "shared_decks_count": {
            "type": "integer",
            "title": "Shared Decks Count",
            "description": "Number of decks shared with user",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "created_at",
          "updated_at",
          "user_id",
          "email"
        ],
        "title": "UserProfileResponse",
        "description": "User profile response with timestamps."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}